backend/requirements.txt
Flask
flask_cors
python-dotenv
requests
openai
whisper
torch